Add b:git_dir to tags even if submodule tags exist
stridx() returns an index if the tags option contains b:git_dir even if
it is as part of a longer path (e.g. a tags file in a submodule).
split() the tags option by ',' and use index() to look for an exact
match of b:git_dir. split() by '[^\\]\zs,' to support escaped commas in
a tags directory.